+++ react/components/index.js
@@ -1,19 +1,21 @@
-import BlurredLoder from './BlurredLoader'
+import BlurredLoader from './BlurredLoader'
 import Carousel from './Carousel'
 import GradientCollapse from './GradientCollapse'
 import ImageResizer from './ImageResizer'
 import ProductDescription from './ProductDescription'
 import ProductName from './ProductName'
+import ProductPrice from './ProductPrice'
 import SKUSelector from './SKUSelector'
 import Video from './Video'
 
 export {
-  BlurredLoder,
+  BlurredLoader,
   Carousel,
   GradientCollapse,
   ImageResizer,
   ProductDescription,
   ProductName,
+  ProductPrice,
   SKUSelector,
   Video,
 }